# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "Campaign": {
        "additionalProperties": false,
        "properties": {
          "active_snapshot_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Snapshot Id"
          },
          "configuration": {
            "$ref": "#/components/schemas/DraftConfiguration"
          },
          "configuration_locked_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Configuration Locked At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Title"
          },
          "has_unpublished_changes": {
            "title": "Has Unpublished Changes",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "public_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Url"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "state": {
            "enum": [
              "draft",
              "published",
              "closed"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "state",
          "revision",
          "configuration",
          "active_snapshot_id",
          "configuration_locked_at",
          "has_unpublished_changes",
          "created_at",
          "public_url"
        ],
        "title": "Campaign",
        "type": "object"
      },
      "CampaignInput": {
        "additionalProperties": false,
        "properties": {
          "configuration": {
            "$ref": "#/components/schemas/DraftConfiguration"
          }
        },
        "required": [
          "configuration"
        ],
        "title": "CampaignInput",
        "type": "object"
      },
      "CampaignPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Campaign"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "CampaignPage",
        "type": "object"
      },
...
      "DateCondition": {
        "additionalProperties": false,
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "kind": {
            "const": "available_by_date",
            "title": "Kind",
            "type": "string"
          },
          "question_id": {
            "format": "uuid",
            "title": "Question Id",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "question_id",
          "date"
        ],
        "title": "DateCondition",
        "type": "object"
      },
...
      "DraftConfiguration": {
        "additionalProperties": false,
        "properties": {
          "deadline": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline"
          },
          "description": {
            "maxLength": 10000,
            "title": "Description",
            "type": "string"
          },
          "domain": {
            "maxLength": 200,
            "title": "Domain",
            "type": "string"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/Question"
            },
            "maxItems": 50,
            "title": "Questions",
            "type": "array"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/Requirement"
            },
            "maxItems": 20,
            "title": "Requirements",
            "type": "array"
          },
          "target_level": {
            "maxLength": 500,
            "title": "Target Level",
            "type": "string"
          },
          "title": {
            "maxLength": 200,
            "title": "Title",
            "type": "string"
          },
          "type": {
            "enum": [
              "training",
              "recruitment"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "domain",
          "description",
          "target_level",
          "deadline",
          "requirements",
          "questions"
        ],
        "title": "DraftConfiguration",
        "type": "object"
      },
...
      "Error": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "Error",
        "type": "object"
      },
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 100,
            "title": "Code",
            "type": "string"
          },
          "details": {
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            },
            "title": "Details",
            "type": "array"
          },
          "message": {
            "maxLength": 2000,
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "maxLength": 100,
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id",
          "details"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 100,
            "title": "Code",
            "type": "string"
          },
          "path": {
            "maxLength": 500,
            "title": "Path",
            "type": "string"
          }
        },
        "required": [
          "path",
          "code"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "Option": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "label"
        ],
        "title": "Option",
        "type": "object"
      },
...
      "Question": {
        "additionalProperties": false,
        "properties": {
          "constraints": {
            "$ref": "#/components/schemas/QuestionConstraints"
          },
          "help": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Help"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "maxLength": 200,
            "title": "Label",
            "type": "string"
          },
          "options": {
            "items": {
              "$ref": "#/components/schemas/Option"
            },
            "maxItems": 20,
            "title": "Options",
            "type": "array"
          },
          "position": {
            "maximum": 49.0,
            "minimum": 0.0,
            "title": "Position",
            "type": "integer"
          },
          "required": {
            "title": "Required",
            "type": "boolean"
          },
          "type": {
            "enum": [
              "short_text",
              "long_text",
              "email",
              "number",
              "date",
              "single_choice",
              "multiple_choice",
              "url",
              "file"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "label",
          "required",
          "position",
          "constraints"
        ],
        "title": "Question",
        "type": "object"
      },
      "QuestionConstraints": {
        "additionalProperties": false,
        "properties": {
          "allowed_media_types": {
            "anyOf": [
              {
                "items": {
                  "enum": [
                    "application/pdf",
                    "image/png",
                    "image/jpeg"
                  ],
                  "type": "string"
                },
                "maxItems": 3,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Media Types"
          },
          "max_bytes": {
            "anyOf": [
              {
                "maximum": 10485760.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Bytes"
          },
          "max_files": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Files"
          },
          "max_items": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Items"
          },
          "max_length": {
            "anyOf": [
              {
                "maximum": 10000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Length"
          },
          "max_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Value"
          },
          "min_items": {
            "anyOf": [
              {
                "maximum": 20.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Items"
          },
          "min_length": {
            "anyOf": [
              {
                "maximum": 10000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Length"
          },
          "min_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Value"
          }
        },
        "title": "QuestionConstraints",
        "type": "object"
      },
...
      "Requirement": {
        "additionalProperties": false,
        "properties": {
          "assessment_mode": {
            "anyOf": [
              {
                "enum": [
                  "automatic",
                  "manual"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Mode"
          },
          "condition_rule": {
            "anyOf": [
              {
                "discriminator": {
                  "mapping": {
                    "available_by_date": "#/components/schemas/DateCondition",
                    "available_slots": "#/components/schemas/SlotsCondition"
                  },
                  "propertyName": "kind"
                },
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/DateCondition"
                  },
                  {
                    "$ref": "#/components/schemas/SlotsCondition"
                  }
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Condition Rule"
          },
          "evaluation_mode": {
            "enum": [
              "qualitative",
              "deterministic"
            ],
            "title": "Evaluation Mode",
            "type": "string"
          },
          "expectation": {
            "maxLength": 2000,
            "title": "Expectation",
            "type": "string"
          },
          "family": {
            "maxLength": 100,
            "title": "Family",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "importance": {
            "enum": [
              "required",
              "desired"
            ],
            "title": "Importance",
            "type": "string"
          },
          "source_question_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 50,
            "title": "Source Question Ids",
            "type": "array"
          }
        },
        "required": [
          "id",
          "family",
          "expectation",
          "importance",
          "evaluation_mode",
          "source_question_ids"
        ],
        "title": "Requirement",
        "type": "object"
      },
...
      "SlotsCondition": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "const": "available_slots",
            "title": "Kind",
            "type": "string"
          },
          "option_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Option Ids",
            "type": "array"
          },
          "question_id": {
            "format": "uuid",
            "title": "Question Id",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "question_id",
          "option_ids"
        ],
        "title": "SlotsCondition",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Talent Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/campaigns": {
      "get": {
        "operationId": "list_campaigns",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Precondition Required"
          }
        },
        "summary": "Listing",
        "tags": [
          "campaigns"
        ]
      },
      "post": {
        "operationId": "create_campaign",
        "parameters": [
          {
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Campaign"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "428": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Precondition Required"
          }
        },
        "summary": "Create",
        "tags": [
          "campaigns"
        ]
      }
    },
...
  }
}